==> notebook/1_user_list.ipynb <==
import sys
import os
import gc
sys.path.append(r"..")

from utils import *
import pandas as pd
import numpy as np

path_build = "../../data/tencent2020/build/"
path_save = "../../data/tencent2020/save/"

pickle_path = "../../data/tencent2020/pickle"
if not os.path.exists(pickle_path):
    os.mkdir(pickle_path) 
pd.set_option('display.max_columns', None)

train_click_log = pd.read_csv(path_save + "train_click_log.csv", encoding="utf-8")
test_click_log = pd.read_csv(path_save + "test_click_log.csv", encoding="utf-8")
train_click_log.shape, test_click_log.shape

train：user_id 范围 (1, 900000)  
test ：user_id 范围 (3000001, 4000000) 

click_log = train_click_log.append(test_click_log)
click_log.sort_values(by=['user_id','time'], ascending=[True,True], inplace=True)
click_log = click_log.reset_index(drop=True)
click_log = click_log[click_log['click_times']<=8] #去除点击次数99.99%外的数据记录
print(click_log.shape)
click_log.head()

## 统计特征生成

# 统计ctr特征
user_ids = pd.read_pickle(f"{pickle_path}/user_ids_relencode.pickle")
# base_features = ['creative_id', 'ad_id']
# dense_features = []
# for fea in base_features:
#     for g in [1]:
#         dense_features.append(f'mean_{fea}_gender_{g}')
#         dense_features.append(f'sum_{fea}_gender_{g}')
#     for a in [1,2,3,4,5,6,7,8,9,10]:
#         dense_features.append(f'mean_{fea}_age_{a}')
#         dense_features.append(f'sum_{fea}_age_{a}')
# user_ids = user_ids.drop(dense_features, axis=1)      
# print(user_ids.columns) 
# user_ids.to_pickle(f"{pickle_path}/user_ids_relencode.pickle")


user = pd.read_csv(train_preliminary_p + "user.csv", encoding='utf-8')
ctr_df = train_click_log.merge(user, how='left', on='user_id')
ctr_df = pd.concat([ctr_df, pd.get_dummies(ctr_df['age'],prefix='age'), pd.get_dummies(ctr_df['gender'],prefix='gender')], axis=1)
del ctr_df['gender_2']
print(ctr_df.head())


gender_p = dict(user['gender'].value_counts() / user_ids.shape[0])
age_p = dict(user['age'].value_counts() / user_ids.shape[0])
print(gender_p, age_p)

ctr_list = ['gender_1', 'age_1', 'age_2', 'age_3', 'age_4', 'age_5', 'age_6', 'age_7', 'age_8', 'age_9', 'age_10']
del user
gc.collect()

#numeric_features = ['creative_id', 'ad_id', 'product_id', 'product_category', 'advertiser_id', 'industry']
numeric_features = ['ad_id', 'product_id', 'product_category', 'advertiser_id', 'industry']

for fea in numeric_features:
    print(fea)
    creative_id_ctr = ctr_df.drop_duplicates(['user_id',fea])

    creative_id_ctr_df = creative_id_ctr.groupby([fea])[ctr_list].mean().add_prefix(f'mean_{fea}_').reset_index()
    creative_id_sum = creative_id_ctr.groupby([fea])[ctr_list].sum().add_prefix(f'sum_{fea}_').reset_index()
    creative_id_size = creative_id_ctr.groupby([fea]).size().reset_index().rename({0:'size'}, axis=1)
    creative_id_ctr_df[f'size'] = creative_id_size['size']
    #贝叶斯平滑
    for f in ctr_list:
        M = creative_id_ctr_df[f'mean_{fea}_{f}'].mean()
        S = creative_id_ctr_df[f'mean_{fea}_{f}'].var()
        alpha =( M * ((M*(1-M))/(S)-1)  )*10
        beta  =( (1-M) * ((M*(1-M))/(S)-1) )*10
        creative_id_ctr_df[f'sum_{fea}_{f}'] = creative_id_sum[f'sum_{fea}_{f}']
        creative_id_ctr_df[f'mean_{fea}_{f}'] = (creative_id_ctr_df[f'sum_{fea}_{f}'] + alpha) / (creative_id_ctr_df[f'size'] + alpha + beta)


    click_log = click_log.merge(creative_id_ctr_df, how='left', on=fea)

    #填充缺失值
    click_log[f'mean_{fea}_gender_1'] = click_log[f'mean_{fea}_gender_1'].fillna(gender_p[1])
    click_log[f'sum_{fea}_gender_1'] = click_log[f'sum_{fea}_gender_1'].fillna(0)
    click_log[f'size'] = click_log['size'].fillna(0)
    for i in range(1,11):
        click_log[f'mean_{fea}_age_{i}'] = click_log[f'mean_{fea}_age_{i}'].fillna(age_p[i])
        click_log[f'sum_{fea}_age_{i}'] = click_log[f'sum_{fea}_age_{i}'].fillna(0)

    creative_id = click_log.groupby(['user_id'])[ [f'mean_{fea}_'+i for i in ctr_list]+[f'sum_{fea}_'+i for i in ctr_list] ].sum().reset_index()
    #user_ids = pd.concat([user_ids, creative_id.drop(['user_id'], axis=1)], axis=1)
    creative_id.to_pickle(f"{pickle_path}/{fea}_ctr_df.pickle")
    del creative_id_sum,creative_id_size,creative_id_ctr_df
    gc.collect()
    #user_ids = user_ids.merge(creative_id, how='left', on='user_id')
#user_ids.head()
#user_ids.to_pickle(f"{pickle_path}/user_ids_relencode.pickle")

user_ids.head()

creative_id = pd.read_pickle(f"{pickle_path}/creative_id_ctr_df.pickle")
ad_id = pd.read_pickle(f"{pickle_path}/ad_id_ctr_df.pickle")
reduce_mem_usage(creative_id)
reduce_mem_usage(ad_id)
user_ids = pd.concat([user_ids, creative_id.drop(['user_id'], axis=1), ad_id.drop(['user_id'], axis=1)], axis=1)
user_ids.to_pickle(f"{pickle_path}/user_ids_ctr.pickle")

#user_ids = pd.concat([user_ids, creative_id.drop(['user_id'], axis=1)], axis=1)
#creative_id2 = pd.read_pickle(f"{pickle_path}/creative_id_ctr_df.pickle")
#user_ids = pd.concat([user_ids, creative_id2.drop(['user_id'], axis=1)], axis=1)
#user_ids.to_pickle(f"{pickle_path}/user_ids_relencode.pickle")

#user_ids.head()
user_ids.columns

user_ids.columns
user_ids.to_pickle(f"{pickle_path}/user_ids_relencode.pickle")
#creative_id.shape

time_df = click_log.groupby(['user_id','time'])['click_times'].sum().reset_index()
time_df = multi_column_LabelEncoder(time_df,columns=['click_times'])
time_df = time_df.groupby('user_id')['click_times'].apply(lambda x: list(x)).reset_index()

ad_df = click_log.groupby(['user_id','time']).size().reset_index()
ad_df.columns = ['user_id','time','id_size']
ad_df = multi_column_LabelEncoder(ad_df,columns=['id_size'])
ad_df = ad_df.groupby('user_id')['id_size'].apply(lambda x: list(x)).reset_index()

user_ids = pd.read_pickle(f"{pickle_path}/user_ids_relencode.pickle")
user_ids['time_clicktimes'] = time_df['click_times']
user_ids['time_creativeids'] = ad_df['id_size']
user_ids.to_pickle(f"{pickle_path}/user_ids_relencode.pickle")
user_ids.head()


#click_log['time_7'] = click_log['time']//7
# click_log['creative_id_times'] = click_log['creative_id']*10000000 + click_log['click_times']
# click_log['ad_id_times'] = click_log['ad_id']*10000000 + click_log['click_times']
# click_log['product_id_times'] = click_log['product_id']*10000000 + click_log['click_times']
# click_log['advertiser_id_times'] = click_log['advertiser_id']*10000000 + click_log['click_times']
# click_log['product_category_times'] = click_log['product_category']*10000000 + click_log['click_times']
# click_log['industry_times'] = click_log['industry']*10000000 + click_log['click_times']


click_log['creative_id_times'] = click_log['creative_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
click_log['ad_id_times'] = click_log['ad_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
click_log['product_id_times'] = click_log['product_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
click_log['advertiser_id_times'] = click_log['advertiser_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
click_log['product_category_times'] = click_log['product_category'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
click_log['industry_times'] = click_log['industry'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')


click_log = multi_column_LabelEncoder(click_log, 
                                      columns=['creative_id','ad_id', 'product_id', 'product_category', 'advertiser_id','industry' ,
                                               'creative_id_times','ad_id_times','product_id_times',
                                               'advertiser_id_times','product_category_times', 'industry_times',
                                               
                                              ])
# click_log = multi_column_LabelEncoderOrder(click_log, 
#                                       columns=['creative_id_times','ad_id_times','product_id_times',
#                                                'advertiser_id_times','product_category_times', 'industry_times',   
#                                                 'creative_id_t','ad_id_t','product_id_t',
#                                                 'advertiser_id_t','product_category_t','industry_t',                                             
#                                               ], max_features=2000000)

print(click_log['creative_id'].nunique())
print(click_log['ad_id'].nunique())
print(click_log['product_id'].nunique())
print(click_log['product_category'].nunique())
print(click_log['advertiser_id'].nunique())
print(click_log['industry'].nunique())
print(click_log['creative_id_times'].nunique())
print(click_log['ad_id_times'].nunique())
print(click_log['product_id_times'].nunique())
print(click_log['advertiser_id_times'].nunique())
print(click_log['product_category_times'].nunique())
print(click_log['industry_times'].nunique())

# print(click_log['creative_id_t'].nunique())
# print(click_log['ad_id_t'].nunique())
# print(click_log['product_id_t'].nunique())
# print(click_log['advertiser_id_t'].nunique())
# print(click_log['product_category_t'].nunique())
# print(click_log['industry_t'].nunique())

user_ids = click_log.groupby(['user_id']).agg({'creative_id':lambda x:list(x),
                                                        'ad_id':lambda x:list(x),
                                                        'product_id':lambda x:list(x),
                                                        'product_category':lambda x:list(x),
                                                        'advertiser_id':lambda x:list(x),
                                                        'industry':lambda x:list(x),   
                                                        'click_times':lambda x:list(x),
                                                         'time':lambda x:list(x),
                                                'creative_id_times':lambda x:list(x),
                                                'ad_id_times':lambda x:list(x),
                                                'product_id_times':lambda x:list(x),
                                                'advertiser_id_times':lambda x:list(x),
                                                'product_category_times':lambda x:list(x),
                                                'industry_times':lambda x:list(x),
#                                                'creative_id_t':lambda x:list(x),
#                                                 'ad_id_t':lambda x:list(x),
#                                                 'product_id_t':lambda x:list(x),
#                                                 'advertiser_id_t':lambda x:list(x),
#                                                 'product_category_t':lambda x:list(x),
#                                                 'industry_t':lambda x:list(x),
                                                       }).reset_index()

user_ids['creative_id_len'] = user_ids['creative_id'].map(lambda x: len(set(x)))
user_ids['ad_id_len'] = user_ids['ad_id'].map(lambda x: len(set(x)))
user_ids['product_id_len'] = user_ids['product_id'].map(lambda x: len(set(x)))
user_ids['product_category_len'] = user_ids['product_category'].map(lambda x: len(set(x)))
user_ids['advertiser_len'] = user_ids['advertiser_id'].map(lambda x: len(set(x)))
user_ids['industry_len'] = user_ids['industry'].map(lambda x: len(set(x)))
user_ids['mean_clicktimes'] = user_ids['click_times'].map(lambda x: np.mean(x))
user_ids['max_clicktimes'] = user_ids['click_times'].map(lambda x: np.max(x))
user_ids['min_clicktimes'] = user_ids['click_times'].map(lambda x: np.min(x))

user_ids['mean_time'] = user_ids['time'].map(lambda x: np.mean(x))
user_ids['max_time'] = user_ids['time'].map(lambda x: np.max(x))
user_ids['min_time'] = user_ids['time'].map(lambda x: np.min(x))

user_ids['time_len'] = user_ids['time'].map(lambda x: len(set(x)))

user_ids.to_pickle(f"{pickle_path}/user_ids_relencode.pickle")
user_ids.head(2)
