==> notebook/3_w2v.ipynb <==
import sys
import os
sys.path.append(r"..")

from utils import *
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression

from sklearn.metrics import accuracy_score,roc_auc_score,f1_score,recall_score
from gensim import models
from scipy import sparse
from scipy.sparse import csr_matrix, coo_matrix 
path_build = "../../data/tencent2020/build/"
path_save = "../../data/tencent2020/save/"
pickle_path = "../../data/tencent2020/pickle"
sub_path = "../../data/tencent2020/sub"

train_preliminary_p = path_build + "train_preliminary/"

if not os.path.exists(pickle_path):
    os.mkdir(pickle_path) 
if not os.path.exists(sub_path):
    os.mkdir(sub_path)

user_ids = pd.read_pickle(f"{pickle_path}/user_ids_relencode.pickle")
user = pd.read_csv(train_preliminary_p + "user.csv", encoding='utf-8')
user_ids = user_ids.merge(user, how='left', on='user_id')
user_ids.head()

w2v_features = [
#     {'name':'creative_id', 'size':128, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'ad_id', 'size':128, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'advertiser_id', 'size':64, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'product_id', 'size':64, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'industry', 'size':32, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'product_category', 'size':16, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'time', 'size':16, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'click_times', 'size':8, 'windows':5, 'min_count':1, 'version':2},
    
#     {'name':'creative_id_times', 'size':128, 'windows':5, 'min_count':1, 'version':2},
#     ##{'name':'ad_id_times', 'size':128, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'product_id_times', 'size':64, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'advertiser_id_times', 'size':64, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'product_category_times', 'size':32, 'windows':5, 'min_count':1, 'version':2},
#     {'name':'industry_times', 'size':32, 'windows':5, 'min_count':1, 'version':2},
    
##     {'name':'creative_id_t', 'size':128, 'windows':5, 'min_count':1, 'version':2},
##     {'name':'ad_id_t', 'size':128, 'windows':5, 'min_count':1, 'version':2},
##     {'name':'product_id_t', 'size':64, 'windows':5, 'min_count':1, 'version':2},
##     {'name':'advertiser_id_t', 'size':64, 'windows':5, 'min_count':1, 'version':2},
##     {'name':'product_category_t', 'size':32, 'windows':5, 'min_count':1, 'version':2},
##     {'name':'industry_t', 'size':32, 'windows':5, 'min_count':1, 'version':2},
    
    {'name':'time_clicktimes', 'size':91, 'windows':10, 'min_count':1, 'version':2,},
    {'name':'time_creativeids', 'size':91, 'windows':10, 'min_count':1, 'version':2 },
    
]
for fea_dict in w2v_features:
    print(fea_dict)
    name = fea_dict['name']
    size = fea_dict['size']
    windows = fea_dict['windows']
    min_count = fea_dict['min_count']
    version = fea_dict['version']
    user_ids[name] = user_ids[name].map(lambda x: [str(i) for i in x])
    w2v = models.Word2Vec(user_ids[name].values, sg=1, size=size, window=windows, workers=6, hs=1, min_count=min_count)
    w2v.wv.save_word2vec_format(f"{path_save}/{name}_w2v_{windows}_{version}.bin",binary=True)
    models.Word2Vec()
    #保存矩阵
    #w2v = models.KeyedVectors.load_word2vec_format(f"{path_save}/{name}_w2v_{size}.bin",binary=True)
    w2v = models.KeyedVectors.load_word2vec_format(f"{path_save}/{name}_w2v_{windows}_{version}.bin",binary=True)
    if 'vocab_size' in fea_dict:
        w2v_vocab_size = fea_dict['vocab_size']
    else:
        w2v_vocab_size = len(w2v.vocab) + 1
        print(w2v_vocab_size)
    w2v_matrix = np.zeros((w2v_vocab_size, w2v.vector_size))
    for w in w2v.vocab:
        w2v_matrix[int(w)] = w2v[w]
    save_pickle(w2v_matrix.astype(np.float16), f"{path_save}/{name}_w2v_matrix2.pkl",)

Word2Vec参数：  
hs ({0, 1}, optional) – 1: 采用hierarchical softmax训练模型; 0: 使用负采样。  
1.skip-gram （训练速度慢，对罕见字有效），CBOW（训练速度快）。一般选择Skip-gram模型；  

2.训练方法：Hierarchical Softmax（对罕见字有利），Negative Sampling（对常见字和低维向量有利）；  

3.欠采样频繁词可以提高结果的准确性和速度（1e-3~1e-5）  

4.Window大小：Skip-gram通常选择10左右，CBOW通常选择5左右。
